Add support to set DefaultRegion on CloudIdentity
When using the SDK with then open source version of OpenStack, we need
to specify a region on every single method call on IComputeProvider,
since OpenStack does not allow to configure default region on the user.
This change adds the option to set a default region on the CloudIdentity
that overrides whatever is retrieved from the server.