Specify the test_script to resolve test failures
By default, all test assemblies are loaded into a single process for execution.
This mode causes a problem since different test assemblies depend on different
versions of the JSON.NET library, and the test runner was not able to load the
required versions side-by-side. The simplest solution to this was explicitly
specifying the test script to run each assembly in its own process.