Retry requests on 401 Unauthorized result
When a token expires, the API returns 401 Unauthorized. At that point
the SDK should request a new token and retry the request.

The new OpenStack.HttpTest class enables unit testing this behavior.
If the Flurl HttpTest class is used, then our special message handler is
not used and this authentication behavior is not exercised in unit tests.
Which should not be an issue, unless the user was explicitly attempting
to test out the retry logic.