Standardize on IChildResource and SetParent
Child resources, such as a ServerVolume or ImageMetadata often
need references back to the parent. It should always be protected
and only exposed through IChildResource. The ApiBuilder is responsible
for setting a parent reference or id.

The type should be reference, and then try casting to a full object
if additional functionality is needed. The child won't always have
a full object available, if it must have one, then the method
should be moved elsewhere.